requirements.txt
aiohttp==3.7.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-countries==6.1.3
django-environ==0.4.5
django-webpack-loader==0.7.0
djangorestframework==3.12.1
environ==1.0
idna==2.10
joblib==0.17.0
maxminddb==2.0.3
multidict==5.0.0
numpy==1.19.4
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.4
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.25.10
xmltodict==0.12.0
yarl==1.6.2
